requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.1.0=pypi_0
accelerate=0.28.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
arxiv-latex-cleaner=1.0.8=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
av=12.1.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bitsandbytes=0.45.0=pypi_0
bleach=6.1.0=pypi_0
blinker=1.7.0=pypi_0
blis=0.7.11=pypi_0
braceexpand=0.1.7=pypi_0
bypy=1.8.5=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.2.0=pypi_0
catalogue=2.0.10=pypi_0
certifi=2022.6.15=pypi_0
cffi=1.17.1=pypi_0
cfgv=3.4.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
colorlog=6.9.0=pypi_0
comm=0.2.2=pypi_0
confection=0.1.4=pypi_0
configargparse=1.7=pypi_0
contexttimer=0.3.3=pypi_0
contourpy=1.2.0=pypi_0
crcmod=1.7=pypi_0
cryptography=44.0.0=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
dash=2.16.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
decorator=5.1.1=pypi_0
decord=0.6.0=pypi_0
diffusers=0.27.2=pypi_0
dill=0.3.8=pypi_0
distlib=0.3.8=pypi_0
einops=0.7.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fairscale=0.4.4=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.1=pypi_0
flask=3.0.2=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.47.2=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
gast=0.5.3=pypi_0
gdown=5.2.0=pypi_0
gin-config=0.5.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
google-auth=2.8.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.3=pypi_0
h5py=3.7.0=pypi_0
huggingface-hub=0.25.0=pypi_0
identify=2.5.35=pypi_0
idna=3.3=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=4.11.4=pypi_0
imutils=0.5.4=pypi_0
iopath=0.1.10=pypi_0
ipython=8.22.2=pypi_0
ipywidgets=8.1.2=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kaggle=1.6.6=pypi_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
langcodes=3.3.0=pypi_0
largestinteriorrectangle=0.2.0=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=14.0.1=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.11.2=pypi_0
llvmlite=0.42.0=pypi_0
markdown=3.3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
more-itertools=10.5.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.10=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
nodeenv=1.8.0=pypi_0
numba=0.59.1=pypi_0
numpy=1.22.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.3.0=pypi_0
open3d=0.18.0=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
opendatasets=0.1.22=pypi_0
openssl=3.0.12=h7f8727e_0
opt-einsum=3.3.0=pypi_0
oss2=2.19.1=pypi_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
platformdirs=4.2.0=pypi_0
plotly=5.20.0=pypi_0
portalocker=2.8.2=pypi_0
pre-commit=3.6.2=pypi_0
preshed=3.0.9=pypi_0
progressbar2=4.4.2=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=3.19.0=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=15.0.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.21.0=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-magic=0.4.27=pypi_0
python-slugify=8.0.4=pypi_0
python-utils=3.8.2=pypi_0
pytorch-fid=0.3.0=pypi_0
pytorch-lightning=2.3.0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.34.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.28.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
requests-toolbelt=1.0.0=pypi_0
retrying=1.3.4=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.8=pypi_0
ruamel-yaml=0.18.6=pypi_0
ruamel-yaml-clib=0.2.12=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
smmap=5.0.1=pypi_0
soupsieve=2.6=pypi_0
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=pypi_0
stack-data=0.6.3=pypi_0
stitching=0.5.3=pypi_0
streamlit=1.32.2=pypi_0
super-image=0.1.7=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.2=pypi_0
tensorflow-addons=0.16.1=pypi_0
tensorflow-estimator=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
text-unidecode=1.3=pypi_0
thinc=8.2.3=pypi_0
threadpoolctl=3.4.0=pypi_0
tifffile=2024.2.12=pypi_0
timm=0.4.9=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.21.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.1=pypi_0
torchaudio=2.2.1=pypi_0
torchmetrics=1.4.0.post0=pypi_0
torchvision=0.17.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
transformers=4.48.0.dev0=pypi_0
triton=2.2.0=pypi_0
typeguard=2.13.3=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=1.26.9=pypi_0
virtualenv=20.25.1=pypi_0
visual-genome=1.1.1=pypi_0
wasabi=1.1.2=pypi_0
watchdog=4.0.0=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.3.4=pypi_0
webdataset=0.2.86=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.10=pypi_0
wrapt=1.14.1=pypi_0
xz=5.4.5=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.13=h5eee18b_0
